+++ extensions/spotify/src/main/java/app/revanced/extension/spotify/misc/privacy/SanitizeSharingLinksPatch.java
@@ -33,10 +33,11 @@ public final class SanitizeSharingLinksPatch {
                 }
             }
 
-            return builder.build().toString();
+            String sanitizedUrl = builder.build().toString();
+            Logger.printInfo(() -> "Sanitized url " + url + " to " + sanitizedUrl);
+            return sanitizedUrl;
         } catch (Exception ex) {
-            Logger.printException(() -> "sanitizeUrl failure", ex);
-
+            Logger.printException(() -> "sanitizeUrl failure with " + url, ex);
             return url;
         }
     }