+++ extensions/youtube/src/main/java/app/revanced/extension/youtube/patches/components/ShortsFilter.java
@@ -218,7 +218,7 @@ public final class ShortsFilter extends Filter {
         //
         suggestedActionsGroupList.addAll(
                 new ByteArrayFilterGroup(
-                        Settings.HIDE_SHORTS_COMMENT_PANEL,
+                        Settings.HIDE_SHORTS_PREVIEW_COMMENT,
                         // Preview comment that can popup while a Short is playing.
                         // Uses no bundled icons, and instead the users profile photo is shown.
                         "shorts-comments-panel"
@@ -266,6 +266,10 @@ public final class ShortsFilter extends Filter {
                 new ByteArrayFilterGroup(
                         Settings.HIDE_SHORTS_GREEN_SCREEN_BUTTON,
                         "greenscreen_temp"
+                ),
+                new ByteArrayFilterGroup(
+                        Settings.HIDE_SHORTS_NEW_POSTS_BUTTON,
+                        "yt_outline_box_pencil"
                 ),
                 new ByteArrayFilterGroup(
                         Settings.HIDE_SHORTS_HASHTAG_BUTTON,