+++ patches/src/main/kotlin/app/revanced/patches/googlephotos/misc/backup/Fingerprints.kt
@@ -1,0 +1,8 @@
+package app.revanced.patches.googlephotos.misc.backup
+
+import app.revanced.patcher.fingerprint
+
+internal val isDCIMFolderBackupControlDisabled = fingerprint {
+    returns("Z")
+    strings("/dcim", "/mars_files/")
+}